requirements.txt
dj-database-url==0.5.0
Django==2.1.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
whitenoise==3.3.1
pip==18.1
setuptools==39.1.0
matplotlib==3.0.1
Pillow==5.3.0
